+++ pkg/dto/history/timer_events.go
@@ -1,10 +1,10 @@
 package history
 
 type TimerCreated struct {
-	TimerID int32
+	TimerID int64
 	FireAt  int64
 }
 
 type TimerFired struct {
-	TimerID int32
+	TimerID int64
 }
